json test

+++ src/main/java/app/standard/Util.java
@@ -1,9 +1,15 @@
 package app.standard;
 
+import app.domain.wiseSaying.WiseSaying;
+
 import java.io.IOException;
 import java.nio.file.*;
 import java.nio.file.attribute.BasicFileAttributes;
+import java.util.Arrays;
+import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 public class Util {
     public static class File {
@@ -34,6 +40,10 @@ public class Util {
 
             Path filePath = Paths.get(file);
 
+            if (filePath.getParent() != null) {
+                createDir(filePath.getParent().toString());
+            }
+
             try {
                 Files.writeString(filePath, content, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
             } catch (IOException e) {
@@ -42,15 +52,19 @@ public class Util {
             }
         }
 
-        public static void delete(String file) {
+        public static boolean delete(String file) {
 
             Path filePath = Paths.get(file);
 
+            if (!Files.exists(filePath)) return false;
+
             try {
                 Files.delete(filePath);
+                return true;
             } catch (IOException e) {
                 System.out.println("파일 삭제 실패");
                 e.printStackTrace();
+                return false;
             }
         }
 
@@ -64,9 +78,11 @@ public class Util {
         }
 
         public static void deleteForce(String path) {
+
             Path folderPath = Paths.get(path);
 
             if (!Files.exists(folderPath)) return;
+
             try {
                 // 디렉토리 및 내용 삭제
                 Files.walkFileTree(folderPath, new SimpleFileVisitor<>() {
@@ -86,29 +102,93 @@ public class Util {
                         return FileVisitResult.CONTINUE;
                     }
                 });
+
                 System.out.println("폴더와 그 안의 내용이 성공적으로 삭제되었습니다.");
             } catch (IOException e) {
                 System.err.println("폴더 삭제 중 오류 발생: " + e.getMessage());
             }
         }
 
+        public static List<Path> getPaths(String dirPathStr) {
 
+            Path dirPath = Paths.get(dirPathStr);
+
+            try {
+                return Files.walk(dirPath)
+                        .filter(Files::isRegularFile)
+                        .toList();
+
+            } catch (Exception e) {
+                System.out.println("파일 목록 가져오기 실패");
+                e.printStackTrace();
+            }
+
+            return List.of();
+        }
     }
 
     public static class Json {
 
         public static String mapToJson(Map<String, Object> map) {
 
-            String tmp = "";
-            for(String key : map.keySet()) {
-                String value = map.get("name").toString();
+            StringBuilder jsonBuilder = new StringBuilder();
 
-                tmp = "{\n" + "    \"%s\" : " + "\"%s\"" + "\n}";
-                tmp = tmp.formatted(key, value);
-                System.out.println(key + " : " + map.get(key));
+            jsonBuilder.append("{\n");
+
+            String str = map.keySet().stream()
+                    .map(k -> map.get(k) instanceof String
+                            ? "    \"%s\" : \"%s\"".formatted(k, map.get(k))
+                            : "    \"%s\" : %s".formatted(k, map.get(k))
+                    ).collect(Collectors.joining(",\n"));
+
+            jsonBuilder.append(str);
+            jsonBuilder.append("\n}");
+
+            return jsonBuilder.toString();
+        }
+
+        public static void writeAsMap(String filePath, Map<String, Object> wiseSayingMap) {
+            String jsonStr = mapToJson(wiseSayingMap);
+            File.write(filePath, jsonStr);
+        }
+
+        public static Map<String, Object> readAsMap(String filePath) {
+            String jsonStr = File.readAsString(filePath);
+
+            if(jsonStr.isEmpty()) {
+                return new LinkedHashMap<>();
             }
 
-            return tmp;
+            return jsonToMap(jsonStr);
+        }
+
+        public static Map<String, Object> jsonToMap(String jsonStr) {
+
+            Map<String, Object> resultMap = new LinkedHashMap<>();
+
+            jsonStr = jsonStr.replaceAll("\\{", "")
+                    .replaceAll("}", "")
+                    .replaceAll("\n", "")
+                    .replaceAll(" : ", ":");
+
+            Arrays.stream(jsonStr.split(","))
+                    .map(p -> p.trim().split(":"))
+                    .forEach(p -> {
+                        String key = p[0].replaceAll("\"", "");
+                        String value = p[1];
+
+                        if(value.startsWith("\"")) {
+                            resultMap.put(key, value.replaceAll("\"", ""));
+                        } else if(value.contains(".")) {
+                            resultMap.put(key, Double.parseDouble(value));
+                        } else if(value.equals("true") || value.equals("false")) {
+                            resultMap.put(key, Boolean.parseBoolean(value));
+                        } else {
+                            resultMap.put(key, Integer.parseInt(value));
+                        }
+                    });
+
+            return resultMap;
         }
     }
 }
